autodev-bot/requirements.txt
python-dotenv
PyGithub
openai>=1.0.0
